data_collector/requirements.txt
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

